+++ src/Handler/HandlerInterface.php
@@ -4,15 +4,15 @@ declare(strict_types=1);
 
 namespace GuzzleLogMiddleware\Handler;
 
-use Exception;
 use GuzzleHttp\TransferStats;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Log\LoggerInterface;
+use Throwable;
 
 /**
  * Classes that will implement this interface are responsible
- * to log the MessageInterface|\Exception|TransferStats that are
+ * to log the MessageInterface|\Throwable|TransferStats that are
  * passed as values.
  *
  * @author George Mponos <gmponos@gmail.com>
@@ -22,9 +22,9 @@ interface HandlerInterface
     public function log(
         LoggerInterface $logger,
         RequestInterface $request,
-        ?ResponseInterface $response,
-        ?Exception $exception,
-        ?TransferStats $stats,
-        array $options
+        ?ResponseInterface $response = null,
+        ?Throwable $exception = null,
+        ?TransferStats $stats = null,
+        array $options = []
     ): void;
 }